requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector search and embeddings
pymilvus==2.3.4
sentence-transformers==2.3.1

# Database clients
pymongo==4.6.1
duckdb==0.9.2
python-tds==1.15.0

# LLM (temporary - will be removed)
openai==1.10.0

# Scientific computing
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.2.1
pycryptodome==3.19.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Visualization
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
streamlit==1.31.0
pandas==2.1.4
